Fix issue with default unsigned char
For some reason on my distro, char defaults to unsigned,
which breaks getopts if it is returned to char.

I fixed it in the code by changing char to int; but I still slightly worry that
there might be an issue somewhere else.

Anyway changing char to int fixes this and is more "correct" as getopt returns int anyway.